+++ tools/src/main/java/org/semanticweb/owlapi/metrics/AverageAssertedNamedSuperclassCount.java
@@ -27,8 +27,7 @@ import org.semanticweb.owlapi.model.OWLOntologyChange;
 import org.semanticweb.owlapi.model.OWLSubClassOfAxiom;
 
 /**
- * @author Matthew Horridge, The University Of Manchester, Bio-Health
- *         Informatics Group
+ * @author Matthew Horridge, The University Of Manchester, Bio-Health Informatics Group
  * @since 2.1.0
  */
 public class AverageAssertedNamedSuperclassCount extends DoubleValuedMetric {
@@ -36,8 +35,7 @@ public class AverageAssertedNamedSuperclassCount extends DoubleValuedMetric {
     /**
      * Instantiates a new average asserted named superclass count.
      * 
-     * @param o
-     *        ontology to use
+     * @param o ontology to use
      */
     public AverageAssertedNamedSuperclassCount(@Nonnull OWLOntology o) {
         super(o);
@@ -61,9 +59,8 @@ public class AverageAssertedNamedSuperclassCount extends DoubleValuedMetric {
                     count++;
                     int prevTotal = total;
                     processedClasses.add(cls);
-                    for (OWLClassExpression desc : sup(
-                            ont.getSubClassAxiomsForSubClass(cls),
-                            OWLClassExpression.class)) {
+                    for (OWLClassExpression desc : sup(ont.getSubClassAxiomsForSubClass(cls),
+                        OWLClassExpression.class)) {
                         if (!desc.isAnonymous()) {
                             total++;
                         }
@@ -74,15 +71,13 @@ public class AverageAssertedNamedSuperclassCount extends DoubleValuedMetric {
                 }
             }
         }
-        return (double) total / count;
+        return Double.valueOf((double) total / count);
     }
 
     @Override
-    protected boolean isMetricInvalidated(
-            List<? extends OWLOntologyChange> changes) {
+    protected boolean isMetricInvalidated(List<? extends OWLOntologyChange> changes) {
         for (OWLOntologyChange chg : changes) {
-            if (chg.isAxiomChange()
-                    && chg.getAxiom() instanceof OWLSubClassOfAxiom) {
+            if (chg.isAxiomChange() && chg.getAxiom() instanceof OWLSubClassOfAxiom) {
                 return true;
             }
         }
